Replace DTO with arrays
